refactor: share the query and header types between the commands
`QueryPair` and `HeaderPair` leave cmd_api.py for a module of their own,
because a second command needs the same two shapes and the details are
easy to get subtly different: where the split happens, what is trimmed,
which names are refused.

The refusal list is the one thing that differs. `api` writes the body
itself and always as JSON, so it owns `content-type` as well; a command
that sends a body as it was given has to leave the type to whoever wrote
it. `HeaderPair` therefore takes the set it refuses.
